+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fallback image on loading error in avatar component
 * Checkbox component
 * Input component
+* Select component
 * Radio button component
 * Snackbar component
 * Progress bar component